\name{dx}
\alias{dx}
\alias{dy}
\alias{dz}
\docType{data}
\title{Elementary forms}
\description{

Objects \code{dx}, \code{dy} and \code{dz} are the three elementary
one-forms on three-dimensional space.  These objects can be generated by
running script \file{vignettes/dx.Rmd}, which includes some further
discussion and technical documentation and creates file \code{dx.rda}
which resides in the \code{data/} directory.

Option \code{usedx} is useful for working with these objects.

}
\usage{data(dx)}
\details{See the vignette for an extended discussion.}
\references{
  \itemize{\item  M. Spivak 1971. _Calculus on manifolds_, Addison-Wesley}
}
\author{Robin K. S. Hankin}
\seealso{\code{\link{d}}}
\examples{


dx
hodge(dx,3)

options(kform_symbolic_print = 'dx')  # shows dx dy dz
dx^dy
hodge(dx,3)

options(kform_symbolic_print = NULL)  # # revert to default


}
\keyword{datasets}
