\name{ex}
\alias{ex}
\alias{ey}
\alias{ez}
\docType{data}
\title{Basis vectors in three-dimensional space}
\description{\loadmathjax

Objects \code{ex}, \code{ey} and \code{ez} are the three elementary
one-forms on three-dimensional space, sometimes denoted
\mjseqn{(e_x,e_y,e_z)}.  These objects can be generated by running
script \file{vignettes/ex.Rmd}, which includes some further discussion
and technical documentation and creates file \file{exeyez.rda} which
resides in the \code{data/} directory.
}
\details{See vignettes \code{dx} and \code{exeyez} for an extended
  discussion; a use-case is given in \code{vector_cross_product}.}
\references{
  \itemize{\item M. Spivak 1971. \emph{Calculus on manifolds},
  Addison-Wesley}
}
\author{Robin K. S. Hankin}
\seealso{\code{\link{d}},\code{\link{print.kform}}}
\examples{

as.function(dx)(ex)


}
\keyword{datasets}
