\name{stokes-package}
\alias{stokes-package}
\docType{package}
\title{
\packageTitle{stokes}
}
\description{

stokes version 1.0-9 is a reduced functionality placeholder pending
upload of a newer spray package to CRAN which uses disordR discipline.

\packageDescription{stokes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stokes}
\packageIndices{stokes}


Generally in the package, arguments that are \eqn{k}-forms are denoted
\code{K}, \eqn{k}-tensors by \code{U}, and spray objects by \code{S}.
Multilinear maps (which may be either \eqn{k}-forms or \eqn{k}-tensors)
are denoted by \code{M}.

}
\author{
\packageAuthor{stokes}

Maintainer: \packageMaintainer{stokes}
}
\references{
  \itemize{
    \item J. H. Hubbard and B. B. Hubbard 2015.  \emph{Vector
      calculus, linear algebra and differential forms: a unified
      approach}.  Ithaca, NY.
    \item M. Spivak 1971. \emph{Calculus on manifolds},
    Addison-Wesley.
  }
}
\keyword{ package }
