\name{Alt}
\alias{Alt}
\title{Alternating multilinear forms}
\description{Converts a k-tensor to alternating form}
\usage{Alt(S)}
\arguments{
  \item{S}{A multilinear form}
}
\details{
details here
}
\value{
Returns an alternating k-tensor
}
\author{Robin K. S. Hankin}
\seealso{\code{\link{help}}}
\examples{

Alt(spray(expand.grid(1:4,1:4),rnorm(16)))

} 