\name{dx}
\alias{dx}
\alias{dy}
\alias{dz}
\docType{data}
\title{Elementary forms}
\description{

Objects \code{dx}, \code{dy} and \code{dz} are the three elementary
one-forms on three-dimensional space.  These objects can be generated by
running script \file{vignettes/dx.Rmd}, which includes some further
discussion and technical documentation and creates file \file{dx.rda}
which resides in the \code{data/} directory.

The default method includes options to print these objects more
intuitively than the default.  Use
\preformatted{options(kform_symbolic_print = "dx")}.

}
\usage{data(dx)}
\details{See the vignette for an extended discussion.}
\references{
  \itemize{\item  M. Spivak 1971. _Calculus on manifolds_, Addison-Wesley}
}
\author{Robin K. S. Hankin}
\seealso{\code{\link{d}},\code{\link{print.kform}}}
\examples{

dx
hodge(dx)
hodge(dx,3)

options(kform_symbolic_print = 'dx')  # shows dx dy dz
dx
dx^dz
hodge(dx,3)

options(kform_symbolic_print = NULL)  # revert to default


}
\keyword{datasets}
