\name{keep}
\alias{keep}
\alias{discard}
\alias{retain}
\title{Keep or drop variables}
\description{Keep or drop variables}
\usage{
keep(K, yes)
discard(K, no)
}
\arguments{
  \item{K}{Object of class \code{kform}}
  \item{yes, no}{Specification of dimensions to either keep (yes) or
    discard (no)}
}
\details{

Given a $k$-form $\omega$, we have
$\omega(v_1,\ldots,v_k)\in\mathbb{R}$, where
$v_1,\ldots,v_k\in\mathbb{R}^n$.  Now, _discarding_ dimension $i$ is
equivalent to asserting (or guaranteeing) that $e_i\cdot v_j=0$ for
$j=1,\ldots,k$.  Alternatively, we may say that $\omega(v_1,\ldots,v_k)$
is independent of $e_i\cdot v_j$ for $j=1,\ldots,k$.  If this is the
case, we may ignore any row in which an $i$ appears.

For $k$-forms, discarding (and its dual, keeping) is carried out in
functions \code{discard()} and \code{keep()}.  Function
\code{keep(omega, yes)} keeps the terms specified and
\code{discard(omega, no)} discards the terms specified.
}
\value{The functions documented here all return a \code{kform} object.}
\author{Robin K. S. Hankin}
\seealso{\code{\link{lose}}}
\examples{

(o <- kform_general(7, 3, seq_len(choose(7, 3))))
keep(o, 1:4)     # keeps only terms with dimensions 1-4
discard(o, 1:2)  # loses any term with "1" or "2" in the index

}
